refactor blur shader and add comments
